file io bug

+++ A2/Sender.java
@@ -47,7 +47,7 @@ public class Sender
 	private static packet[] create_packets_from_file() throws Exception {
 
 		RandomAccessFile file = new RandomAccessFile(file_name, "r");
-		byte[] file_bytes = new byte[file.length()];
+		byte[] file_bytes = new byte[(int) file.length()];
 		file.readFully(file_bytes);
 		file.close();
 		for (int i = 0; i < file_bytes.length; i++) {
